+++ frontend/app/(teacher)/teacher/(authenticated)/colleagues/[colleagueId]/page.tsx
@@ -1,10 +1,10 @@
 import { ColleagueDetail } from "@/features/ColleagueDetail";
 
 type Props = {
-    params: Promise<{colleagueId: string}>;
+  params: Promise<{ colleagueId: string }>;
 };
 
 export default async function ColleaguePage({ params }: Props) {
-    const { colleagueId } = await params;
-    return <ColleagueDetail colleagueId={Number(colleagueId)} />;
+  const { colleagueId } = await params;
+  return <ColleagueDetail colleagueId={Number(colleagueId)} />;
 }
