+++ frontend/app/api/teacher/colleagues/[colleagueId]/route.ts
@@ -2,27 +2,30 @@ import { RailsUnauthorizedError } from "@/libs/server/rails/railsError";
 import { railsFetch } from "@/libs/server/rails/railsFetch";
 import { type NextRequest, NextResponse } from "next/server";
 
-export async function GET(_: NextRequest, { params }: {params: Promise<{colleagueId: string}> }) {
-    const { colleagueId } = await params;
+export async function GET(
+  _: NextRequest,
+  { params }: { params: Promise<{ colleagueId: string }> },
+) {
+  const { colleagueId } = await params;
 
-    try {
-        const { status, data, setCookie } = await railsFetch(
-            `/api/v1/teacher/colleagues/${colleagueId}`,
-        );
+  try {
+    const { status, data, setCookie } = await railsFetch(
+      `/api/v1/teacher/colleagues/${colleagueId}`,
+    );
 
-        const res = NextResponse.json(data, { status })
+    const res = NextResponse.json(data, { status });
 
-        if (setCookie) res.headers.set("set-cookie", setCookie);
+    if (setCookie) res.headers.set("set-cookie", setCookie);
 
-        return res;
-    } catch (error) {
-        if (error instanceof RailsUnauthorizedError) {
-            return NextResponse.json({ message: "UNAUTHORIZED" }, { status: 401 });
-        }
-
-        return NextResponse.json(
-            { message: "INTERNAL_SERVER_ERROR" },
-            { status: 500 },
-        );
+    return res;
+  } catch (error) {
+    if (error instanceof RailsUnauthorizedError) {
+      return NextResponse.json({ message: "UNAUTHORIZED" }, { status: 401 });
     }
+
+    return NextResponse.json(
+      { message: "INTERNAL_SERVER_ERROR" },
+      { status: 500 },
+    );
+  }
 }
